Added file handers for drag and drop from desktop, less improvements, cleanup and generalization

+++ Ucp.class.php
@@ -52,15 +52,17 @@ class Ucp implements BMO {
 		$path = FreePBX::create()->Config->get_conf_setting('AMPWEBROOT');
 		$location = $path.'/ucp';
 		foreach($modules as $module) {
-			$rawname = trim($module['rawname']);
-			if(file_exists($path.'/admin/modules/'.$rawname.'/ucp') && file_exists($path.'/admin/modules/'.$rawname.'/ucp/'.ucfirst($rawname).".class.php")) {
-				if($module['status'] == MODULE_STATUS_ENABLED) {
-					if(!file_exists($location."/modules/".ucfirst($rawname))) {
-						symlink($path.'/admin/modules/'.$rawname.'/ucp',$location.'/modules/'.ucfirst($rawname));
-					}
-				} else {
-					if(file_exists($location."/modules/".ucfirst($rawname)) && is_link($location."/modules/".ucfirst($rawname))) {
-						unlink($location."/modules/".ucfirst($rawname));
+			if(isset($module['rawname'])) {
+				$rawname = trim($module['rawname']);
+				if(file_exists($path.'/admin/modules/'.$rawname.'/ucp') && file_exists($path.'/admin/modules/'.$rawname.'/ucp/'.ucfirst($rawname).".class.php")) {
+					if($module['status'] == MODULE_STATUS_ENABLED) {
+						if(!file_exists($location."/modules/".ucfirst($rawname))) {
+							symlink($path.'/admin/modules/'.$rawname.'/ucp',$location.'/modules/'.ucfirst($rawname));
+						}
+					} else {
+						if(file_exists($location."/modules/".ucfirst($rawname)) && is_link($location."/modules/".ucfirst($rawname))) {
+							unlink($location."/modules/".ucfirst($rawname));
+						}
 					}
 				}
 			}
